requirements.txt
h5py==3.2.1
numpy==1.20.1
Pillow==9.4.0
playsound==1.3.0
scipy==1.6.2
torch==1.9.0a0+2ecb2c7
torchvision==0.9.0a0
